Move sharing menu folder
